Use numpy for bulk pixel data conversion if available
This brings back the Numpy approach to converting pixel data for the SPI
interface from the original Adafruit_Python_ILI9341 library. The use of
Numpy instead of the previous double loop results in a performance boost
of a factor >10 when testing on a Raspberry Pi Zero.

When Numpy is not available at runtime, the previous double loop is
used.

Fixes #46