Allow image drawing in subscreen areas
This extends the `image` method with two additional arguments for the
image origin on the screen. The previous exact size check was adapted
to ensure that the image drawn at the supplied origin does not exceed
the display dimensions.

The changes allow to update a smaller part of the screen with an
image which is especially benefitial on lower end hardware. Some
example test results from a Rasberry Pi Zero:

- Draw fullscreen image (240x320): ~0.3s
- Draw half-screen image (240x160): ~0.2s
- Draw single line of text (240x22): ~0.06s

Fixes: #48