:books: update usage example for 1.14" TFT users
The usage example may only work for 2.2", 2.4", 2.8", 3.2" TFT ( which I could not verify ). With 1.14" TFT and raspberry pi 4, the wiring are different, chipset is different. 

At the first try-out, as a novice user, I was doubting my soldering skills. Why does not it work? Blank screen! It turns out that wrong chipset driver was used. 

After updating its chipset driver, it is still blank! I was doubting my connections and my soldering skills again! It turns out that the CS_PIN, DC_PIN are not the same as bigger TFTs.

After all those adjustments, it worked.